Move app to separate dir
