+++ .golangci.yml
@@ -22,8 +22,6 @@ linters-settings:
       - wrapperFunc
   gocyclo:
     min-complexity: 15
-  goimports:
-    local-prefixes: github.com/petuhovskiy/dist-comp-hw
   golint:
     min-confidence: 0
   gomnd:
