+++ app/repos/psql/products.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/jackc/pgx/v4"
 
-	"github.com/petuhovskiy/dist-comp-hw/modeldb"
+	"app/modeldb"
 )
 
 type Products struct {
