+++ app/routers/router.go
@@ -9,11 +9,11 @@ import (
 	"github.com/go-chi/chi/middleware"
 	"github.com/go-chi/render"
 
-	"github.com/petuhovskiy/dist-comp-hw/handlers"
+	"app/handlers"
 
 	httpSwagger "github.com/swaggo/http-swagger"
 
-	_ "github.com/petuhovskiy/dist-comp-hw/docs" // docs is generated by Swag CLI, you have to import it.
+	_ "app/docs" // docs is generated by Swag CLI, you have to import it.
 )
 
 func CreateRouter(productsV1 *handlers.Products) chi.Router {
