+++ docker-compose.yml
@@ -9,7 +9,7 @@ services:
     restart: unless-stopped
 
   app:
-    build: .
+    build: ./app
     ports:
       - 8080:8080
     environment:
